Move parser combinators (list/set/kv) to Data.Parser
